Working!
We not have Castlevania 1, MegaMan, DuckTales and more!!!

+++ src/coral/cartridge/mapper/types.rs
@@ -1,8 +1,8 @@
 pub trait MapperT {
-    fn cpu_r_mapt(&mut self, address : u16) -> u16;
-    fn cpu_w_mapt(&mut self, address : u16, byte : u8) -> Option<u16>;
-    fn ppu_r_mapt(&mut self, address : u16) -> u16;
-    fn ppu_w_mapt(&mut self, address : u16, byte : u8) -> Option<u16>;
+    fn cpu_r_mapt(&mut self, address : u16) -> usize;
+    fn cpu_w_mapt(&mut self, address : u16, byte : u8) -> Option<usize>;
+    fn ppu_r_mapt(&mut self, address : u16) -> usize;
+    fn ppu_w_mapt(&mut self, address : u16, byte : u8) -> Option<usize>;
     fn clone_self(&self) -> Box<dyn MapperT>;
     fn reset(&mut self);
 }
@@ -12,16 +12,16 @@ pub trait MapperT {
 pub struct Mapper(pub Box<dyn MapperT>);
 
 impl Mapper {
-    pub fn cpu_r_map(&mut self, address : u16) -> u16{
+    pub fn cpu_r_map(&mut self, address : u16) -> usize{
         self.0.cpu_r_mapt(address)
     }
-    pub fn cpu_w_map(&mut self, address : u16, byte : u8) -> Option<u16>{
+    pub fn cpu_w_map(&mut self, address : u16, byte : u8) -> Option<usize>{
         self.0.cpu_w_mapt(address, byte)
     }
-    pub fn ppu_r_map(&mut self, address : u16) -> u16{
+    pub fn ppu_r_map(&mut self, address : u16) -> usize{
         self.0.ppu_r_mapt(address)
     }
-    pub fn ppu_w_map(&mut self, address : u16, byte : u8) -> Option<u16> {
+    pub fn ppu_w_map(&mut self, address : u16, byte : u8) -> Option<usize> {
         self.0.ppu_w_mapt(address, byte)
     }
     pub fn reset(&mut self){
